% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDC.R
\docType{package}
\name{SDCdragon}
\alias{SDCdragon}
\title{Statistical Data Control. Data Research, Access, Governance Network.}
\description{
A tool for checking how much information is disclosed when
reporting summary statistics
}
